% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{dimorder}
\alias{dimorder}
\title{dimorder dummy}
\description{
dimorder dummy
}
\section{Dimension ordering}{

Note the unexpected ordering of the dimensions. We use this convention in this package because R is a column-major language.
}

