% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{extractMatchMatrices}
\alias{extractMatchMatrices}
\title{Extract match matrices from imported table}
\usage{
extractMatchMatrices(marketData)
}
\arguments{
\item{marketData}{The return value of \code{importCommon}.}
}
\value{
A list of match matrices.
}
\description{
Extract match matrices from imported table
}
\section{Match matrix structure}{

Let \code{mIdx} index a market. Each \code{matchMatrix} is an array of
dimension \code{(noD[mIdx], noU[mIdx])}. The element indexed by
\code{[dIdx, uIdx]} is \code{1} if the triple \code{(mIdx, uIdx, dIdx)}
matches and \code{0} otherwise.
}

\section{Old code}{

To port code using the previous version, replace
\code{matchMatrix[[m]][[u]][d]} by \code{matchMatrices[[m]][d, u]}.
}

\section{Dimension ordering}{

Note the unexpected ordering of the dimensions. We use this convention in
this package because R is a column-major language.
}

\keyword{internal}
