% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{generateAssignmentMatrix}
\alias{generateAssignmentMatrix}
\title{Generate assignment matrix}
\usage{
generateAssignmentMatrix(payoffMatrix, quotaU, quotaD)
}
\arguments{
\item{payoffMatrix}{An array of payoffs for all upstream-downstream pairs in
this market. Note that the first dimension indexes the downstreams. See the
\code{payoff.R} file for more details.}

\item{quotaU}{A vector of length \code{numU}, containing integers in the
range \code{0:numD}. Its \code{j}-th element is the maximum allowed number
of matches for the corresponding upstream. Alternatively, it can be a
single non-negative integer. In this case, the quota vector is this value,
repeated the appropriate number of times.}

\item{quotaD}{As \code{quotaU}, but for downstreams.}
}
\value{
An array of dimension \code{(numD, numU)}. Its element indexed by
\code{[dIdx, uIdx]} is \code{1} if that downstream-upstream pair matches,
and \code{0} otherwise.
}
\description{
Finds the optimal matching in a market for the given payoffs and quotas.
}
\details{
The optimal matching maximizes the total payoff.
}
