% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import matched market data}
\usage{
import(filename)
}
\arguments{
\item{filename}{Absolute or relative path to the file. See also the
parameters to \code{\link[data.table]{data.table::fread}}.}
}
\value{
A list with members:
\tabular{ll}{
\code{$header}             \tab A character vector of the headers of the
table.\cr
\code{$noM}                \tab The number of markets.\cr
\code{$noU}, \code{$noD}   \tab Vectors of size \code{$noM}, whose
\code{m}-th element is the number of upstreams and downstreams
respectively in the \code{m}-th market.\cr
\code{$noAttr}             \tab The number of distance attributes.\cr
\code{$distanceMatrices}   \tab A list of arrays of distance values, one
for each market. See the appropriate section for their definition.\cr
\code{$matchMatrices}      \tab A list of arrays of zeros or ones
describing matches, one for each market. See the appropriate section for
their definition.\cr
\code{$mate}               \tab A list of data.table objects describing
matches, one for each market. See the appropriate section for their
definition.\cr
}
Members \code{$matchMatrices} and \code{$mate} provide the same information,
expressed in different ways.
}
\description{
Reads a CSV file containing data for matched markets.
}
\section{File structure}{


The file must be a delimiter-separated file with a header. It must contain
the following fields:
\tabular{ll}{
\code{Market} \tab The market index.\cr
\code{UpStream} \tab The upstream index.\cr
\code{DownStream} \tab The downstream index.\cr
\code{Match} \tab \code{1} if this triple matches, \code{0} otherwise.
}
It must also contain at least one field with a name starting with
\code{Distance}. These fields containg distance attribute values. The order
they appear in, and not their full name, specifies their actual order.

Indices should have consecutive values, starting from \code{1}. Distance
attribute values should be numerical.

Each row should correspond to a unique triple of market, upstream, and
downstream indices.
}

\section{Distance matrix structure}{

Let \code{mIdx} index a market. Each \code{distanceMatrix} is an array (technically not a matrix) of dimension \code{(noAttr, noD[mIdx], noU[mIdx])}. The element indexed by \code{[i, dIdx, uIdx]} gives the \code{i}-th distance attribute value for the triple \code{(mIdx, uIdx, dIdx)}.
}

\section{Match matrix structure}{

Let \code{mIdx} index a market. Each \code{matchMatrix} is an array of dimension \code{(noD[mIdx], noU[mIdx])}. The element indexed by \code{[dIdx, uIdx]} is \code{1} if the triple \code{(mIdx, uIdx, dIdx)} matches and \code{0} otherwise.
}

\section{Mate table structure}{

Let \code{mIdx} index a market. Each mate table is a \code{data.table} object
with fields \code{UpStream} and \code{DownMates}. Each upstream in the market
corresponds to a row, whose \code{DownMates} field contains a vector of the
downstream indices which the upstream is matched to. The rows are ordered
consecutively based on the upstream index \code{UpStream}.
}

